chore: clippy fixes, miri run

+++ src/tokenizer/parse_quotes.rs
@@ -1,9 +1,9 @@
-pub(super) fn rs_do_quote_bs(bytes_s: &[u8], quote: &mut i32) -> String {
+pub(super) fn rs_do_quote_bs(bytes_s: &[u8], quote: &mut u8) -> String {
 	let mut tmp = String::new();
 	for &byte in bytes_s {
-		match (*quote as u8, byte) {
-			(0, b'\'' | b'"') => *quote = byte as i32,
-			(_, q) if q == *quote as u8 => {
+		match (*quote, byte) {
+			(0, b'\'' | b'"') => *quote = byte,
+			(_, q) if q == *quote => {
 				*quote = 0;
 			}
 			_ => {
